feat: ID type 변경

+++ event-service/src/main/java/com/example/eventservice/domain/repository/interaction/InteractionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface InteractionRepository extends JpaRepository<Interaction, Long> {
+public interface InteractionRepository extends JpaRepository<Interaction, Integer> {
 
 
     @Query("delete from Interaction i where i.culturalEvent.id = :culturalEventId and i.userId = :userId and i.likeStar = :likeStar")
